Add set orientation of aircraft to the "attitude" of the tv remote
In a our first trials we found out that the tv remotes attitude is not correctly set. For this reason we guessed the attitude by calculating the euler angles which lead to the measured gravity. This is only a guess since psi - a rotation around the z axis - cannot be measured by only one gravity sensor.